==> scripts/1-littlesis-graph-exploration.ipynb <==
import networkx as nx
from pathlib import Path
from utils import read_json
import pickle

# Load jsons

entities = read_json(Path("../data/littlesis-entities.json"))
relationships = read_json(Path("../data/littlesis-relationships.json"))

# Helper functions

def print_graph_info(G: nx.Graph):
    print("Number of nodes: ", G.number_of_nodes())
    print("Number of edges: ", G.number_of_edges())
    print()
    print("Number of connected components: ", nx.number_connected_components(G))
    connected_components = sorted(nx.connected_components(G), key=len, reverse=True)
    three_largest_components = connected_components[0:3]
    for i, component in enumerate(three_largest_components):
        print(
            f"{i+1}. Size of the connected component: {(len(component)/G.number_of_nodes())*100:.2f}%"
        )

    print()

    # top five nodes based on pagerank
    pr = nx.pagerank(G)
    top_ten_pagerank = sorted(pr, key=pr.get, reverse=True)[:10]
    print("Top 10 nodes based on pagerank:")
    for i, node in enumerate(top_ten_pagerank):
        print(
            f"#{i+1}\t: nodeid ({node}), pagerank ({pr[node]:.6f}) -> {G.nodes[node]['name']} ({G.nodes[node]['primary_ext']})"
        )

# Original littlesis graph

# Build the graph with all entities and relationships
G = nx.Graph()
for entity in entities:
    G.add_node(
        entity["id"],
        id=entity["attributes"]["id"],
        name=entity["attributes"]["name"],
        primary_ext=entity["attributes"]["primary_ext"],
    )
for relationship in relationships:
    G.add_edge(
        relationship["attributes"]["entity1_id"],
        relationship["attributes"]["entity2_id"],
        category_id=relationship["attributes"]["category_id"],
        amount=relationship["attributes"]["amount"],
    )

print("Graph with all entities and relationships:")
print_graph_info(G)

# Only inspecting the biggest connected component

# Create a subgrah with the largest connected component
connected_components = sorted(nx.connected_components(G), key=len, reverse=True)
largest_component: set = connected_components[0]
G_largest = G.subgraph(largest_component)
print("Created the full graph that has the following information:")
print_graph_info(G_largest)

# Save the largest connected component as a pickle

pickle.dump(G_largest, open("../data/littlesis-largest-component.pickle", "wb"))
