==> scripts/99-visualizations.ipynb <==
# Visualizations and figures for the paper

import pandas as pd
import plotly.graph_objects as go
from plotly.subplots import make_subplots
from plotly.graph_objs import Layout

# Voteview ideology scores

voteview = pd.read_csv("../data/voteview-HS117_members.csv", index_col="icpsr")

# Create a new column for marker colors based on 'party_code'
voteview["marker_color"] = voteview["party_code"].apply(
    lambda x: "red" if x == 100 else "blue" if x == 200 else "gray"
)

senate = voteview[voteview["chamber"] == "Senate"]
senate_democrats = senate[senate["party_code"] == 100]
senate_republicans = senate[senate["party_code"] == 200]
house = voteview[voteview["chamber"] == "House"]
house_democrats = house[house["party_code"] == 100]
house_republicans = house[house["party_code"] == 200]


marker_size = 10


fig = make_subplots(
    rows=2,
    cols=1,
    shared_xaxes=True,
    # vertical_spacing=0.9,
    subplot_titles=("Senate", "House"),
)

fig.add_trace(
    go.Scatter(
        x=senate_democrats["nominate_dim1"],
        mode="markers",
        marker_color=senate_democrats["marker_color"],
        name="Democrats",
    ),
    row=1,
    col=1,
)

fig.add_trace(
    go.Scatter(
        x=senate_republicans["nominate_dim1"],
        mode="markers",
        marker_color=senate_republicans["marker_color"],
        name="Republicans",
    ),
    row=1,
    col=1,
)

fig.add_trace(
    go.Scatter(
        x=house_democrats["nominate_dim1"],
        mode="markers",
        marker_color=house_democrats["marker_color"],
        name="Democrats",
    ),
    row=2,
    col=1,
)

fig.add_trace(
    go.Scatter(
        x=house_republicans["nominate_dim1"],
        mode="markers",
        marker_color=house_republicans["marker_color"],
        name="Republicans",
    ),
    row=2,
    col=1,
)

fig.update_yaxes(
    # showgrid=False,
    zerolinecolor="black",
    showticklabels=False,
)
fig.update_xaxes(range=[-1, 1])
fig.update_layout(
    showlegend=False,
    paper_bgcolor="rgba(0,0,0,0)",
    plot_bgcolor="rgba(0,0,0,0)",
    autosize=False,
    width=600,
    height=300,
    margin=dict(l=0, r=0, b=0, t=20, pad=4),
)
# save to png file
fig.write_image("../paper-figures/congress-nominate-dim1.png")
# fig.show()

# Donor ideology predictions in one dimension

# The ids of the politicians in the NVD dataset, stems from the littlesis ids
from utils import read_json
from typing import List

politician_info = read_json("../data/politicians_117.json")

data = {
    # Meta
    "marker_color": [],
    "nominate_dim1": [],
    "chamber": [],
    # Embedding from NVD TSNE
    "tsne-1": [],
    "tsne-2-0": [],
    "tsne-2-1": [],
    # Embedding from NVD PCA
    "pca-1": [],
    "pca-2-0": [],
    "pca-2-1": [],
    # Embedding from NVD ISOMAP
    "isomap-1": [],
    "isomap-2-0": [],
    "isomap-2-1": [],
}
for info in politician_info:
    try:
        if info["voteview_info"]["party_code"] == 100:
            marker_color = "red"
        elif info["voteview_info"]["party_code"] == 200:
            marker_color = "blue"
        else:
            marker_color = "gray"

        chamber: str = info["voteview_info"]["chamber"]
        nominate_dim1: float = info["voteview_info"]["nominate_dim1"]

        # Embedding from NVD TSNE
        TSNE_1: float = info["TSNE-1"][0]
        TSNE_2_0: float = info["TSNE-2"][0]
        TSNE_2_1: float = info["TSNE-2"][1]

        # Embedding from NVD PCA
        PCA_1: float = info["PCA-1"][0]
        PCA_2_0: float = info["PCA-2"][0]
        PCA_2_1: float = info["PCA-2"][1]

        # Embedding from NVD ISOMAP
        ISOMAP_1: float = info["Isomap-1"][0]
        ISOMAP_2_0: float = info["Isomap-2"][0]
        ISOMAP_2_1: float = info["Isomap-2"][1]

        # ADD VALUES
        data["marker_color"].append(marker_color)
        data["nominate_dim1"].append(nominate_dim1)
        data["chamber"].append(chamber)
        data["tsne-1"].append(TSNE_1)
        data["tsne-2-0"].append(TSNE_2_0)
        data["tsne-2-1"].append(TSNE_2_1)
        data["pca-1"].append(PCA_1)
        data["pca-2-0"].append(PCA_2_0)
        data["pca-2-1"].append(PCA_2_1)
        data["isomap-1"].append(ISOMAP_1)
        data["isomap-2-0"].append(ISOMAP_2_0)
        data["isomap-2-1"].append(ISOMAP_2_1)

    except KeyError as e:
        print(e)

assert len(data["tsne-1"]) == len(data["nominate_dim1"])
assert len(data["tsne-1"]) == len(data["chamber"])

df = pd.DataFrame(data)

senate = df[df["chamber"] == "Senate"]
senate_democrats = senate[senate["marker_color"] == "red"]
senate_republicans = senate[senate["marker_color"] == "blue"]
house = df[df["chamber"] == "House"]
house_democrats = house[house["marker_color"] == "red"]
house_republicans = house[house["marker_color"] == "blue"]

df.describe()

GENERATE_PNG = False

marker_size = 10

for dimension_key in ["tsne-1", "pca-1", "isomap-1"]:
    print(df[dimension_key].min(), df[dimension_key].max())

    fig = make_subplots(
        rows=2,
        cols=1,
        shared_xaxes=True,
        subplot_titles=(f"Senate-{dimension_key}", f"House-{dimension_key}"),
    )

    fig.add_trace(
        go.Scatter(
            x=senate_democrats[dimension_key],
            mode="markers",
            marker_color=senate_democrats["marker_color"],
            name="Democrats",
        ),
        row=1,
        col=1,
    )

    fig.add_trace(
        go.Scatter(
            x=senate_republicans[dimension_key],
            mode="markers",
            marker_color=senate_republicans["marker_color"],
            name="Republicans",
        ),
        row=1,
        col=1,
    )

    fig.add_trace(
        go.Scatter(
            x=house_democrats[dimension_key],
            mode="markers",
            marker_color=house_democrats["marker_color"],
            name="Democrats",
        ),
        row=2,
        col=1,
    )

    fig.add_trace(
        go.Scatter(
            x=house_republicans[dimension_key],
            mode="markers",
            marker_color=house_republicans["marker_color"],
            name="Republicans",
        ),
        row=2,
        col=1,
    )

    fig.update_yaxes(
        # showgrid=False,
        zerolinecolor="black",
        showticklabels=False,
    )
    fig.update_xaxes(range=[-1, 1])
    fig.update_layout(
        showlegend=False,
        paper_bgcolor="rgba(0,0,0,0)",
        plot_bgcolor="rgba(0,0,0,0)",
        autosize=False,
        width=600,
        height=300,
        margin=dict(l=0, r=0, b=0, t=20, pad=4),
    )

    if dimension_key == "tsne-1" and GENERATE_PNG:
        fig.write_image(f"../paper-figures/congress-embedding-{dimension_key}.png")

    fig.show()

# Donor ideology predictions in two dimensions

GENERATE_PNG = False

marker_size = 10

for dimension_key_0, dimension_key_1 in [
    ("tsne-2-0", "tsne-2-1"),
    ("pca-2-0", "pca-2-1"),
    ("isomap-2-0", "isomap-2-1"),
]:
    print(dimension_key_0)
    fig = make_subplots(
        rows=1,
        cols=2,
        shared_xaxes=True,
        # vertical_spacing=0.9,
        subplot_titles=(
            f"Senate",
            f"House",
        ),
    )

    fig.add_trace(
        go.Scatter(
            x=senate_democrats[dimension_key_0],
            y=senate_democrats[dimension_key_1],
            mode="markers",
            marker_color=senate_democrats["marker_color"],
            name="Democrats",
        ),
        row=1,
        col=1,
    )

    fig.add_trace(
        go.Scatter(
            x=senate_republicans[dimension_key_0],
            y=senate_republicans[dimension_key_1],
            mode="markers",
            marker_color=senate_republicans["marker_color"],
            name="Republicans",
        ),
        row=1,
        col=1,
    )

    fig.add_trace(
        go.Scatter(
            x=house_democrats[dimension_key_0],
            y=house_democrats[dimension_key_1],
            mode="markers",
            marker_color=house_democrats["marker_color"],
            name="Democrats",
        ),
        row=1,
        col=2,
    )

    fig.add_trace(
        go.Scatter(
            x=house_republicans[dimension_key_0],
            y=house_republicans[dimension_key_1],
            mode="markers",
            marker_color=house_republicans["marker_color"],
            name="Republicans",
        ),
        row=1,
        col=2,
    )

    fig.update_xaxes(range=[-1, 1])
    fig.update_layout(
        showlegend=False,
        paper_bgcolor="rgba(0,0,0,0)",
        plot_bgcolor="rgba(0,0,0,0)",
        autosize=False,
        width=600,
        height=300,
        margin=dict(l=0, r=0, b=0, t=20, pad=4),
    )

    if dimension_key_0 == "tsne-2-0" and GENERATE_PNG:
        fig.write_image(
            f"../paper-figures/congress-embedding-{'-'.join(dimension_key_0.split('-')[:-1])}.png"
        )

    fig.show()

# Summary statistics of donations

## Divided into republicans and democrates, what is the total amount of money donated to each party? What is the distribution of donations?

# data/donation_to_politician.json is mapped by littlesis id

from utils import read_json
from typing import List

donation_info = read_json("../data/donation_to_politician.json")
politician_info = read_json("../data/politicians_117.json")

for littlesis_politician_id, info in donation_info.items():
    for politician in politician_info:
        assert isinstance(politician, dict)
        assert isinstance(politician["ids"], dict)
        try:
            assert isinstance(politician["ids"]["littlesis"], int)
        except KeyError:
            print("FUCK!")
        assert isinstance(politician["ids"]["icpsr"], int)
        assert isinstance(littlesis_politician_id, str)

        try:
            if politician["ids"]["littlesis"] == littlesis_politician_id:
                icpsr: int = politician["ids"]["icpsr"]
                break
            else:
                icpsr = None

        except KeyError:
            print(politician)
            exit()

    if icpsr is None:
        continue
    else:
        sum_donation: float = sum([float(donation) for donation in info.values()])
        print(littlesis_politician_id, icpsr, sum_donation)

## Who are the top 10 givers for each party?

# Politician case study

Consider 4 four politicians. One from each extreme of the spectrum and two from the middle.

voteview = pd.read_csv("../data/voteview-HS117_members.csv", index_col="icpsr")
senate = voteview[voteview["chamber"] == "Senate"]
house = voteview[voteview["chamber"] == "House"]

# The two most extreme senators
senate.sort_values(by="nominate_dim1", inplace=True)
extremes = senate.iloc[[0, -1]]

# The most extreme republican and democrat
extreme_republican = extremes.loc[extremes.party_code == 200]
extreme_democrat = extremes.loc[extremes.party_code == 100]

# The most moderate republican and democrat
moderate_republican = (
    senate.loc[senate.party_code == 200]
    .sort_values(by="nominate_dim1", key=lambda x: abs(x))
    .iloc[[0]]
)

moderate_democrat = (
    senate.loc[senate.party_code == 100]
    .sort_values(by="nominate_dim1", key=lambda x: abs(x))
    .iloc[[0]]
)

for politician in [
    extreme_republican,
    extreme_democrat,
    moderate_republican,
    moderate_democrat,
]:
    party = "republican" if politician["party_code"].item() == 200 else "democrate"
    name = politician["bioname"].item()
    score = politician["nominate_dim1"].item()
    icpsr = politician.index.item()
    print(f"{party}\ticpsr:{icpsr}\tnominate_dim1 of {score}\t{name}")

## Cosine similarity between the vectors

## Shortest paths between them

# Calculate the shortest path between all four senators
import pickle

G_largest = pickle.load(open("../data/littlesis-largest-component.pickle", "rb"))

# Calculate the shortest path between all four senators
from networkx import shortest_path

shortest_path(extreme_republican.index[0], extreme_democrat.index[0], G_largest)
